+++ lib/output/azure_queue_storage.go
@@ -33,7 +33,7 @@ In order to set the ` + "`queue_name`" + ` you can use function interpolations d
 				"ttl", "The TTL of each individual message as a duration string. Defaults to 0, meaning no retention period is set",
 				"60s", "5m", "36h",
 			).IsInterpolated(),
-			docs.FieldCommon("max_in_flight", "The maximum number of messages to have in flight at a given time. Increase this to improve throughput."),
+			docs.FieldCommon("max_in_flight", "The maximum number of messages to have in flight at a given time. Increase this to improve throughput.").AtVersion("3.45.0"),
 			batch.FieldSpec(),
 		},
 		Categories: []Category{